+++ components/freshchatConfig.js
@@ -11,8 +11,9 @@ class FreshchatConfig {
         this.themeName = null;
         this.stringsBundle = null;
         this.teamMemberInfoVisible = true;
-        this.cameraCaptureEnabled = false;
-        this.gallerySelectionEnabled = false;
+        this.cameraCaptureEnabled = true;
+        this.gallerySelectionEnabled = true;
+        this.responseExpectationEnabled = true;
         this.showNotificationBanner = true; //iOS only
         this.notificationSoundEnabled = true; //iOS only